move exec_instruction to its own file
